ml-service/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML libraries
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
scipy==1.11.4
pillow==10.1.0

# Computer Vision
albumentations==1.3.1
imgaug==0.4.0

# Data processing
pandas==2.1.3
scikit-learn==1.3.2

# Image processing
Pillow==10.1.0
imageio==2.33.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: GPU support
# tensorflow-gpu==2.15.0
# torch==2.1.1+cu118
# torchvision==0.16.1+cu118 